Add installable entrypoint
This adds an entrypoint that is installed as an executable alongside the
package. The 'tests' folder is no longer installed with the rest of the
code.

The entrypoint can optionally be passed an argument to specify the path
to the configuration file to use instead of the default. The advantage
of this is twofold:

 - no messing in /etc
 - custom config paths make sharing a single config across machines
   easier