fix negative frequency updating for moving gratings
This fixes the weird behaviour observed with the moving grating stimulus
when providing a negative frequency.
It should now update properly on screen. Since we were using a buffer
of twice the period for the grating image, I have moved the default
position to be one period up and we now check for the top of the image
rectangle to reset it when it passes the origin threshold.