delay plugin loading until app start
Currently, importing `visiomode.core` leads to automatically loading all
the plugins (i.e., `devices`, `protocols`, and `stimuli`). Although this
makes sense in the context of the full application and is the expected
use case for the end user, this is a bit cumbersome when trying to
create minimal tests.
To follow the spirit of the previous commit, putting plugin loading
behind a flag in the main application initialisation makes it possible
to import the various modules without incurring the burden of loading
the plugins.

Also note that importing `visiomode.core` is not really a choice as
`visiomode.__init__` itself imports `visiomode.core`, making it a
requirement to load the plugins regardless of what imports are done
by the user. This is especially cumbersome when developing tests for
parts of the application that do not make use of the plugins (e.g.,
`visiomode.config`) but which are forced to incur the import cost.