add input and reward address attributes to config
This allows the config to dictate what ports are presented as options on
the main web interface.
The previous approach only listed serial ports
that were automatically detected. This commit provides the user with the
option of providing custom ports through the config, which is useful
when debugging on machines that do not have serial ports available.