remove unnecessary Javascript and HTML element
The `btn` was unused and the "dropdown" div made the download button
take up the whole table cell, meaning that the dropdown menu would be
much bigger than necessary. Additionally, it made it so clicks on the
dropdown were not registered as click events by the table. It is not
noticeable by the user at the moment, but future commits will add a
summary modal to the history and clicking the last column would lead to
no action being detected.