add modal dialogue with session summary when clicking on session history
It is now possible to click on a row in the history table and get a
summary of the session information. The Javascript implementation makes
use of the new GET API and only fetches the session of interest.
If for some reason, the session file is no longer available when trying
to populate the modal elements (e.g., the file has been removed but the
page has not been refreshed since), empty strings are used instead.
Similarly, if some fields (e.g., experimenter name) are not present in
the file, the elements are populated with empty strings.