add functionality for deleting individual session data from the history
The summary modal now has two buttons: cancel and delete. Cancel simply
closes the modal, while delete pops up another modal. That modal informs
the user deleting the session data is irreversible. The user then needs
to hover the confirmation button for one second before being able to
click it.

On click, the button sends a POST request to the `HistoryAPI`, asking it
to remove the session file from disk.

With these precautions, accidental deletions should essentially be
impossible.