reimplement models using pydantic
Models that we previously defined inside of `models.py` are now
implemented using pydantic. This means they are validated at time of
instantiation. As a result, it is possible to simply retrieve
information form a POST request and feed it directly into the model
constructor without needing to do manual validation.

Since the models are now purely data containers, they are no longer
responsible for storing themselves (e.g., `Animal`). A future commit
will implement a `Database` class that is responsible for managing
a database of the model type it was instantiated with.

Note the web API and the models are not currently connected but will
be once the refactor is a bit further along.

Also note, the current `_models` directory is there to avoid name
clashing with `models.py`. It will be changed in future commits once
 the refactor is complete.