add `database` module for models management
The new `database` module introduces a new class `Database` that is used
to store a single type of model. For example, when the application
requires a database for animals, it can simply call
`database.get_database(Animal)`. This will return the appropriate
database object for that model, and any subsequent calls will return
the same object.

These databases use a so-called "uniqueness-key" to ensure there are no
duplicates added to the database.
The functionality mirrors what was available from the `Animal` class of
`models.py` with a few more bells and whistles because why not.