attempt to fix config serialisation error
Previously, we tried serialising a
`serial.list_ports_common.ListPortInfo` which is not serialisable.
This attempts to fix it by only keeping the device name rather than the
whole object.
The fix is tentative since I cannot test this on my own machine as I've
been unable to fake a COM port.